msb/src/requirements.txt
pycti
PyYAML
urllib3
python-dateutil
xmltodict